Ignore the template text that Kindle puts in when 10% limit is exceeded
Amazon has imposed a limit on the portion of the book which can be highlighted at 10% of the total
text. This limit can be exceeded on my device for especially large books (such as Tolstoy's Anna
Karenina.) So, instead of putting the actual text of the highlight, the My Clippings.txt file has
this template text alongwith the locations of the highlight. These locations can be used by the
Kindle to show a highlight when one is reading the book on the Kindle, however, the text can not be
exported using the clippings file.

The workaround for this is to use bookcision.js instead.